twitter-to-pubsub/twitter-to-pubsub-image/requirements.txt
google-cloud-pubsub
tweepy